# src/main/resources/swagger/api.yml
openapi: "3.0.0"
info:
  title: "Hexagonal Architecture"
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: local server

paths:
  /api/orders:
    post:
      summary: "주문 등록"
      operationId: createOrder
      description: "주문을 등록 한다"
      tags:
        - Order Management
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

    get:
      summary: "주문 목록 조회"
      operationId: listOrders
      description: "주문 목록을 조회한다"
      tags:
        - Order Management
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/OrderStatus'

  /api/orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      summary: "주문 상세정보 조회"
      operationId: getOrder
      description: "주문 상세정보를 조회한다"
      tags:
        - Order Management
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

components:

  parameters:
    PageSize:
      in: query
      name: size
      description: 페이지 사이즈(결과수)
      required: false
      schema:
        type: integer
        format: int32
        default: 20
    Page:
      in: query
      name: page
      description: 현재 페이지
      required: false
      schema:
        type: integer
        format: int32
        default: 1
    OrderId:
      in: path
      name: orderId
      description: 주문 ID
      required: true
      schema:
        type: string
        format: uuid
        example: f604e2e7-7e22-4a82-bc27-bc6b255b6976
    OrderStatus:
      name: status
      description: 주문 상태
      in: query
      schema:
        $ref: '#/components/schemas/OrderStatus'

  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string

    Order:
      properties:
        id:
          description: 주문 ID
          type: string
          format: uuid
          example: 71f3764e-b4df-4a20-9194-1682beee0991
        orderNumber:
          description: 주문 번호
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'

    OrderStatus:
      description: 주문 상태
      type: string
      enum:
        - SUBMITTED
        - READY
        - COMPLETED
        - CANCELLED

    CreateOrderRequest:
      properties:
        exceldata:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderRequestExcel'

    CreateOrderRequestExcel:
      properties:
        orderNumber:
          description: 주문번호
          type: string
          example: 01012345678

    ListOrderResponse:
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        page:
          $ref: '#/components/schemas/PageMetadata'

    PageMetadata:
      properties:
        size:
          title: 페이지 당 갯수
          type: integer
          format: int32
          example: 20
        totalElements:
          title: 총 갯수
          type: integer
          format: int64
          example: 1130
        totalPages:
          title: 총 page 수
          type: integer
          format: int32
          example: 57
        number:
          title: page 번호
          type: integer
          format: int32
          example: 1

  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFoundResponse:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalServerErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'